Initial Commit of Suspension Forces Calculator
The suspension forces script calculates the load going through each member of the suspension on each corner of the car, for a number of defined load cases.

Suspension_points_better (terrible name lmao) is a quick little thing to pull the points from the spreadsheet and make it more workable in matlab.

matrixBuildingExample explains the calculations, and is meant to serve as methematical documentation/education on how the link forces are calculated. Probably need something similar to explain how the load cases are translated into tire forces, there's a little more than meets the eye there, and some key assumptions