PI testing
messing around with the controller. everytime you want to tune you have to drag the sliders to get the response you want, so im not sure how to automate. you have to use discrete for discrete and continuous for continuous, you cannot just change the option inside the PID block itself