Script to generate QP matrix C function
MATLAB file which creates a c function file calculating the Q and p matrix elements. I'm doing this automatically because the equations are huge and easy to mess up. TODO: There are some optimizations to be done to make the embedded computation much quicker, such as calculating half the elements and then copying values because the matrix is symmetrical, remove pow(), precompute constant terms, remove common sub expressions, etc